# api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/work-logs": {
      "post": {
        "operationId": "WorkLogController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkLogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Work log created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkLog"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Project or task not found"
          }
        },
        "summary": "Create a new work log entry",
        "tags": [
          "work-logs"
        ]
      },
      "get": {
        "operationId": "WorkLogController_findAll",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Work logs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkLog"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all work logs for a user",
        "tags": [
          "work-logs"
        ]
      }
    },
    "/work-logs/project/{projectId}/summary": {
      "get": {
        "operationId": "WorkLogController_getProjectSummary",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project summary retrieved successfully"
          }
        },
        "summary": "Get project work summary",
        "tags": [
          "work-logs"
        ]
      }
    },
    "/work-logs/{id}": {
      "get": {
        "operationId": "WorkLogController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Work log retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkLog"
                }
              }
            }
          },
          "404": {
            "description": "Work log not found"
          }
        },
        "summary": "Get a specific work log by ID",
        "tags": [
          "work-logs"
        ]
      },
      "patch": {
        "operationId": "WorkLogController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkLogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Work log updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkLog"
                }
              }
            }
          },
          "404": {
            "description": "Work log not found"
          }
        },
        "summary": "Update a work log",
        "tags": [
          "work-logs"
        ]
      },
      "delete": {
        "operationId": "WorkLogController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Work log deleted successfully"
          },
          "404": {
            "description": "Work log not found"
          }
        },
        "summary": "Delete a work log",
        "tags": [
          "work-logs"
        ]
      }
    },
    "/work-logs/reports/monthly": {
      "get": {
        "operationId": "WorkLogController_getMonthlyReport",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "description": "Year (YYYY)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": true,
            "in": "query",
            "description": "Month (1-12)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Monthly report retrieved successfully"
          }
        },
        "summary": "Get monthly work log report",
        "tags": [
          "work-logs"
        ]
      }
    },
    "/work-logs/reports/export/excel": {
      "get": {
        "operationId": "WorkLogController_exportToExcel",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "description": "Year (YYYY)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": true,
            "in": "query",
            "description": "Month (1-12)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Excel file generated successfully"
          }
        },
        "summary": "Export monthly work logs to Excel",
        "tags": [
          "work-logs"
        ]
      }
    },
    "/work-logs/reports/export/pdf": {
      "get": {
        "operationId": "WorkLogController_exportToPdf",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "description": "Year (YYYY)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": true,
            "in": "query",
            "description": "Month (1-12)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF file generated successfully"
          }
        },
        "summary": "Export monthly work logs to PDF",
        "tags": [
          "work-logs"
        ]
      }
    },
...
  },
  "info": {
    "title": "Work Schedule API",
    "description": "API para gerenciamento de agendas de trabalho",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateWorkLogDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Date of work",
            "example": "2024-01-15"
          },
          "startTime": {
            "type": "string",
            "description": "Start time of work",
            "example": "2024-01-15T09:00:00Z"
          },
          "endTime": {
            "type": "string",
            "description": "End time of work",
            "example": "2024-01-15T17:00:00Z"
          },
          "hoursWorked": {
            "type": "number",
            "description": "Hours worked",
            "example": 8
          },
          "description": {
            "type": "string",
            "description": "Description of work done",
            "example": "Implemented user authentication module"
          },
          "taskId": {
            "type": "string",
            "description": "Task ID if work is related to a specific task"
          },
          "projectId": {
            "type": "string",
            "description": "Project ID"
          },
          "userId": {
            "type": "string",
            "description": "User ID"
          }
        },
        "required": [
          "date",
          "startTime",
          "hoursWorked",
          "description",
          "projectId",
          "userId"
        ]
      },
      "WorkLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Work log ID"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Date of work"
          },
          "startTime": {
            "format": "date-time",
            "type": "string",
            "description": "Start time of work"
          },
          "endTime": {
            "format": "date-time",
            "type": "string",
            "description": "End time of work"
          },
          "hoursWorked": {
            "type": "number",
            "description": "Hours worked"
          },
          "description": {
            "type": "string",
            "description": "Description of work done"
          },
          "taskId": {
            "type": "string",
            "description": "Task ID if work is related to a specific task"
          },
          "projectId": {
            "type": "string",
            "description": "Project ID"
          },
          "userId": {
            "type": "string",
            "description": "User ID"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp"
          },
          "calculatedCost": {
            "type": "number",
            "description": "Calculated cost based on project hourly rate"
          }
        },
        "required": [
          "id",
          "date",
          "startTime",
          "hoursWorked",
          "description",
          "projectId",
          "userId",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateWorkLogDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}